Implementados mensajes de error, arreglo de leaks y organizacion de funciones en proceso
