+++ src/build_ins/pwd.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   pwd.c                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: miniore <miniore@student.42.fr>            +#+  +:+       +#+        */
+/*   By: porellan <porellan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/14 12:17:56 by frlorenz          #+#    #+#             */
-/*   Updated: 2025/08/06 20:58:58 by miniore          ###   ########.fr       */
+/*   Updated: 2025/08/13 18:48:21 by porellan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ void    pwd(t_backpack *backpack)
     path = getcwd(NULL, 0); // getcwd asigna memoria automáticamente????????
     if (path == NULL)
     {
-        ft_put_syserr(backpack, "getcwd"); //Mirar como sale el error!!!!!!
+        ft_put_syserr(backpack, "Minichelita"); //Mirar como sale el error!!!!!!
         return ;
     }
     printf("%s\n", path);
