+++ src/error/error.c
@@ -6,7 +6,7 @@
 /*   By: porellan <porellan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/06 21:02:07 by miniore           #+#    #+#             */
-/*   Updated: 2025/08/12 21:20:29 by porellan         ###   ########.fr       */
+/*   Updated: 2025/08/13 18:44:53 by porellan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,6 +33,6 @@ void    ft_put_pererr(t_backpack *backpack, char *err, int n)
 {
     ft_putstr_fd(err, 2);
     backpack->exit_status = n;
-    backpack->commands_nb = 0;
+    //backpack->commands_nb = 0;
     backpack->err_flag = 1;
 }
