+++ src/execution/exec_redir.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   exec_redir.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: frlorenz <frlorenz@student.42.fr>          +#+  +:+       +#+        */
+/*   By: porellan <porellan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/08 21:05:27 by miniore           #+#    #+#             */
-/*   Updated: 2025/08/12 13:04:55 by frlorenz         ###   ########.fr       */
+/*   Updated: 2025/08/13 19:13:02 by porellan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,7 +63,7 @@ static void	ft_redir_heredoc(t_redir *redirection, int exec)
 	char *input;
 
 	if (pipe(pipe_fd) == -1)
-        exit_error();
+        exit_error();  //cambiarrrrr!!!!!
     rl_clear_history();
     while(1)
     {
