+++ src/execution/executor.c
@@ -6,7 +6,7 @@
 /*   By: porellan <porellan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/14 12:43:18 by frlorenz          #+#    #+#             */
-/*   Updated: 2025/08/12 21:23:41 by porellan         ###   ########.fr       */
+/*   Updated: 2025/08/13 19:38:54 by porellan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ void executor(t_backpack *backpack, char **envp, t_env *path)
     else if(ft_strcmp(backpack->commands_lst[backpack->n].command, "cd")  == 0)
         cd(backpack);
     else if(ft_strcmp(backpack->commands_lst[backpack->n].command, "exit")  == 0)
-        ft_exit(backpack); // habria que mirar de liberar.......
+        ft_exit(backpack);
     else if(ft_strcmp(backpack->commands_lst[backpack->n].command, "env")  == 0)
         ft_env(backpack, 1);
     else if(ft_strcmp(backpack->commands_lst[backpack->n].command, "export")  == 0)
