+++ src/parse/syntax_parse.c
@@ -6,83 +6,14 @@
 /*   By: porellan <porellan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/29 20:39:38 by porellan          #+#    #+#             */
-/*   Updated: 2025/08/12 18:53:13 by porellan         ###   ########.fr       */
+/*   Updated: 2025/08/13 20:31:31 by porellan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-int ft_redir_syntax(char *input)
+static int  ft_pipe_checker(char *input, size_t len)
 {
-    size_t len;
-
-    len = 0;
-    while(input[len])
-    {
-        while(input[len] && ft_is_redirct(input[len]))
-        {
-            if(!ft_is_quotes(input[len]) || !ft_is_dquotes(input[len]))
-                len = ft_ignore_qargs(input, len);
-            len++;
-        }
-        if(!ft_is_redirct(input[len]))
-        {
-            len++;
-            if((!ft_is_redirct(input[len]) && !ft_is_redirct(input[len + 1])) ||
-                    (!ft_is_redirct(input[len]) && (input[len - 1] != input[len])))
-                return(EXIT_FAILURE);
-            len++;
-            while (input[len] && !ft_is_space(input[len]))                
-                len++;
-            if(!input[len] || !ft_is_redirct(input[len]) || input[len] == '|')
-                return (EXIT_FAILURE);
-        }
-    }
-    return(EXIT_SUCCESS);
-}
-
-int	ft_quotes_syntax(char *input)
-{
-	int	flag;
-	int	i;
-
-	flag = 0;
-	i = 0;
-	while(input[i] != '\0')
-	{
-		if(!ft_is_quotes(input[i]))
-        {
-            flag = 1;
-            i++;
-            while(ft_is_quotes(input[i]) && input[i] != '\0')
-                i++;
-            if(!ft_is_quotes(input[i]))
-                flag = 0;
-        }
-		if(!ft_is_dquotes(input[i]))
-        {
-            flag = 1;
-            i++;
-            while(ft_is_dquotes(input[i]) && input[i] != '\0')
-                i++;
-            if(!ft_is_dquotes(input[i]))
-                flag = 0;
-        }
-		if(flag == 1)
-			return(EXIT_FAILURE);
-		i++;
-	}
-	return(EXIT_SUCCESS);
-}
-
-int ft_pipe_syntax(char *input)
-{
-    size_t len = 0;
-
-    while(!ft_is_space(input[len]))
-        len++;
-    if(input[len] == '|')
-        return(EXIT_FAILURE);
     while(input[len] != '\0')
     {
         if(input[len] == '|' && input[len + 1] != '|')
@@ -106,7 +37,61 @@ int ft_pipe_syntax(char *input)
                 return(EXIT_FAILURE);
         }
     }
+    return(EXIT_SUCCESS);
+}
+
+static int ft_pipe_syntax(char *input)
+{
+    size_t len = 0;
+
+    while(!ft_is_space(input[len]))
+        len++;
+    if(input[len] == '|')
+        return(EXIT_FAILURE);
+    if(ft_pipe_checker(input, len))
+        return(EXIT_FAILURE);
     return (EXIT_SUCCESS);
+}
+
+static int ft_flag_check(char *input, int i, int flag)
+{
+    while(input[i] != '\0')
+    {
+        if(!ft_is_quotes(input[i]))
+        {
+            flag = 1;
+            i++;
+            while(ft_is_quotes(input[i]) && input[i] != '\0')
+                i++;
+            if(!ft_is_quotes(input[i]))
+                flag = 0;
+        }
+        if(!ft_is_dquotes(input[i]))
+        {
+            flag = 1;
+            i++;
+            while(ft_is_dquotes(input[i]) && input[i] != '\0')
+                i++;
+            if(!ft_is_dquotes(input[i]))
+                flag = 0;
+        }
+        if(flag == 1)
+            return(EXIT_FAILURE);
+        i++;
+    }
+    return(EXIT_SUCCESS);
+}
+
+static int	ft_quotes_syntax(char *input)
+{
+	int	flag;
+	int	i;
+
+	flag = 0;
+	i = 0;
+	if(ft_flag_check(input, i, flag))
+        return(EXIT_FAILURE);
+	return(EXIT_SUCCESS);
 }
 
 int ft_syntax_parse(t_backpack *backpack, char *input)
